clean up test structure, make support module add xvfb for gui tests in travis
