+++ .travis.yml
@@ -6,6 +6,11 @@ python:
   - 3.5
   - 3.4
 
+before_script:
+  - "export DISPLAY=:99.0"
+  - "sh -e /etc/init.d/xvfb start"
+  - sleep 3 # give xvfb some time to start
+
 script:
   python -m unittest discover tests --buffer
