+++ tests/test_split.py
@@ -11,18 +11,13 @@ try:
 except ImportError:
     from mock import patch
 
+from support.stdout_helper import get_stdout
+
 class TestSplit(unittest.TestCase):
     def __init__(self, methodName = "runTest"):
         unittest.TestCase.__init__(self, methodName)
         self.filename = "./tests/fixtures/qdc1.xml"
         self.outputdir = "./tests/fixtures/out"
-
-    # https://stackoverflow.com/questions/4219717/how-to-assert-output-with-nosetest-unittest-in-python
-    # output helper (hides pdb output too)
-    def get_stdout(self):
-        if not hasattr(sys.stdout, "getvalue"):
-            self.fail("need to run in buffered mode")
-        return sys.stdout.getvalue().strip()
 
     def test_process_file_not_found(self):
         with self.assertRaises(EnvironmentError) as context:
@@ -34,7 +29,7 @@ class TestSplit(unittest.TestCase):
         args = argparse.Namespace(filename = self.filename, outputdir = self.outputdir)
         with patch('argparse.ArgumentParser.parse_args', return_value = args):
             self.assertEqual((self.filename, self.outputdir), Split().setup_params())
-            output = self.get_stdout()
+            output = get_stdout()
             self.assertTrue("Processing input file " + self.filename in output)
             self.assertTrue("Creating output xml files in " + self.outputdir in output)
 
@@ -45,13 +40,8 @@ class TestSplit(unittest.TestCase):
             with patch('argparse.ArgumentParser.parse_args',return_value = args):
                 with self.assertRaises(SystemExit):
                     Split().setup_params()
-                    output = self.get_stdout()
+                    output = get_stdout()
                     self.assertTrue("Invalid input or output location" in output)
-
-    def test_setup_params_gui_args(self):
-        # import pdb; pdb.set_trace()
-        # gui = True
-        pass
     
 
 if __name__ == '__main__':
